Seem to have fixed the problem with repeated loops
X12::Loop#parse was iterating through the nodes and parsing them without
regard for the required number of instances. For example, if the loop
had a GE segment and an AA segment and both were required, the loop
would consider a sole GE segment a success. Parse would return the AA
segment and the rest of the document would fail to parse.

Started keeping track of the status of each sub loop/segment and require
all to pass unless there was a limit marking them as optional.